On a morning from a Bogart movie, in a country where they turned back time, You go strolling through the crowd like Peter Lorre contemplating a crime. She comes out of the sun in a silk dress running like a watercolor in the rain. Don't bother asking for explanations, she'll just tell you that she came In the Year of the Cat.
She doesn't give you time for questions, as she locks up your arm in hers,
And you follow 'till your sense of which direction completely disappears.
By the blue-tiled walls near the market stall there's a hidden door she
    leads you to.
These days, she say, I feel my life just like a river running through
The Year of the Cat.

Well, she looks at you so coolly,
And her eyes shine like the moon in the sea.
She comes in incense and patchouli,
So you take her to find what's waiting inside
The Year of the Cat.

Well, morning comes and you're still with her, but the bus and the tourists
    are gone,
And you've thrown away your choice and lost your ticket, so you have to stay on.
But the drum-beat strains of the night remain in the rhythm of the new-born day.
You know some time you're bound to leave her, but for now you're going to stay
In the Year of the Cat.
		-- Al Stewart, "Year of the Cat"